seperate admin cog from gamble cog
