+++ nyanbit/__main__.py
@@ -1,4 +1,3 @@
-import asyncio
 import os
 from dotenv import load_dotenv
